python/flask-webservices-labs/flask-spyne-fc20-labs/examples-fc20-labs/cambio-moedas-wsdl/client_cambio-moedas-wsdl.py
from suds.client import Client as SudsClient
import requests

print("request wsdl")
url = 'http://www.webservicex.com/CurrencyConvertor.asmx?wsdl'
print(url)
client = SudsClient(url=url, cache=None)
resp_request = client.service.ConversionRate(FromCurrency='BRL', ToCurrency='GBP')
print(resp_request)

