python/flask-webservices-labs/flask-spyne-fc20-labs/examples-fc20-labs/cambio-moedas-wsdl/requirements-spyne.txt
#pip freeze do flask spyne instalado com pip
attrs==17.3.0
Automat==0.6.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
Django==1.11.7
Flask==0.12.2
Flask-Spyne==0.3.1
http==0.2
hupper==1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
msgpack-python==0.4.8
PasteDeploy==1.5.2
plaster==1.0
plaster-pastedeploy==0.4.2
pyparsing==2.2.0
pyramid==1.9.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
repoze.lru==0.7
requests==2.18.4
rpc==0.1.3
sec-wall==1.2
simplejson==3.13.2
six==1.11.0
soap==0.0.1rc0
spyne==2.12.14
SQLAlchemy==1.1.15
suds-jurko==0.6
thrift==0.8.0
translationstring==1.3
Twisted==17.9.0
urllib3==1.22
venusian==1.1.0
WebOb==1.7.4
Werkzeug==0.12.2
zope.deprecation==4.3.0
zope.interface==4.4.3
